backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.3.2
underthesea==9.4.0